proper error handling

+++ src/main/java/com/ramailo/service/GenericMiddleware.java
@@ -7,6 +7,7 @@ import javax.inject.Inject;
 import javax.json.JsonObject;
 
 import com.ramailo.RequestInfo;
+import com.ramailo.exception.ResourceNotFoundException;
 import com.ramailo.meta.Action;
 
 public class GenericMiddleware {
@@ -23,7 +24,7 @@ public class GenericMiddleware {
 			return genericService.invokeAction(request, action.get(), null);
 		}
 
-		if (request.getFirstPathParam() != null) {
+		if (request.getFirstPathParam() != null && request.getPathParams().size() == 2) {
 			Object result = genericService.findById(request);
 
 			return result;
@@ -62,7 +63,11 @@ public class GenericMiddleware {
 			genericService.invokeAction(request, action.get(), null);
 			return;
 		}
+		if (request.getPathParams().size() == 2) {
+			genericService.remove(request);
+			return;
+		}
 
-		genericService.remove(request);
+		throw new ResourceNotFoundException();
 	}
 }
